+++ src/Nosto.php
@@ -62,6 +62,10 @@ class Nosto
     const DEFAULT_NOSTO_API_BASE_URL = 'https://api.nosto.com';
     const DEFAULT_NOSTO_IFRAME_ORIGIN_REGEXP = '(https:\/\/(.*)\.hub\.nosto\.com)|(https:\/\/my\.nosto\.com)'; //codingStandardsIgnoreLine
 
+    const URL_PARAM_MESSAGE_TYPE = 'message_type';
+    const URL_PARAM_MESSAGE_CODE = 'message_code';
+    const URL_PARAM_MESSAGE_TEXT = 'message_text';
+
     /**
      * Return environment variable.
      *
@@ -114,12 +118,15 @@ class Nosto
         $message = '';
         $jsonResponse = $response->getJsonResult();
 
-        if (
-            isset($jsonResponse->type)
+        $errors = self::parseErrorsFromResponse($response);
+        if (isset($jsonResponse->type)
             && isset($jsonResponse->message)
         ) {
             if (isset($jsonResponse->message)) {
-                $message .= '. ' . $jsonResponse->message;
+                $message .= $jsonResponse->message;
+            }
+            if (!empty($errors)) {
+                $message .= ' | ' . $errors;
             }
             throw new ApiResponseException(
                 $message,
@@ -130,7 +137,10 @@ class Nosto
             );
         } else {
             if ($response->getMessage()) {
-                $message .= '. ' . $response->getMessage();
+                $message .= $response->getMessage();
+            }
+            if (!empty($errors)) {
+                $message .= ' | ' . $errors;
             }
             throw new HttpResponseException(
                 $message,
@@ -140,5 +150,31 @@ class Nosto
                 $response
             );
         }
+    }
+
+    /**
+     * Parses errors from HttpResponse
+     * @param HttpResponse $response
+     * @return string
+     */
+    public static function parseErrorsFromResponse(HttpResponse $response)
+    {
+        $json = $response->getJsonResult();
+        $errorStr = '';
+        if (isset($json->errors)
+            && is_array($json->errors)
+            && !empty($json->errors)
+        ) {
+            foreach ($json->errors as $stdClassError) {
+                if (isset($stdClassError->errors)) {
+                    $errorStr .= $stdClassError->errors;
+                }
+                if (isset($stdClassError->product_id)) {
+                    $errorStr .= sprintf('(product #%s)', $stdClassError->product_id);
+                }
+            }
+        }
+
+        return $errorStr;
     }
 }