+++ src/Object/AbstractPerson.php
@@ -76,9 +76,14 @@ abstract class AbstractPerson extends AbstractObject implements PersonInterface
      */
     private $country;
 
+    /**
+     * @var boolean the opt-in status for the person
+     */
+    private $marketingPermission;
+
     public function __construct()
     {
-        // Dummy
+        $this->setMarketingPermission(false);
     }
 
     /**
@@ -187,5 +192,30 @@ abstract class AbstractPerson extends AbstractObject implements PersonInterface
     public function setCountry($country)
     {
         $this->country = $country;
+    }
+
+    /**
+     * @return bool
+     */
+    public function getMarketingPermission()
+    {
+        return $this->marketingPermission;
+    }
+
+    /**
+     * @param bool $marketingPermission
+     */
+    public function setMarketingPermission($marketingPermission)
+    {
+        $this->marketingPermission = (bool)$marketingPermission;
+    }
+
+    /**
+     * @param bool $optedIn
+     * @deprecated will be removed in near future, use setMarketingPermission instead
+     */
+    public function setOptedIn($optedIn)
+    {
+        $this->setMarketingPermission($optedIn);
     }
 }