+++ src/Object/MarkupableString.php
@@ -59,9 +59,21 @@ class MarkupableString extends AbstractObject implements MarkupableInterface
         $this->markupKey = $markupKey;
     }
 
+    /**
+     * @return string the value
+     */
     public function getValue()
     {
         return $this->value;
+    }
+
+    /**
+     * Setter for the value
+     * @param string $value
+     */
+    public function setValue($value)
+    {
+        $this->value = $value;
     }
 
     /**