+++ src/Object/Signup/Signup.php
@@ -110,6 +110,7 @@ class Signup extends Settings implements SignupInterface
         $this->addApiToken(Token::API_SSO);
         $this->addApiToken(Token::API_EXCHANGE_RATES);
         $this->addApiToken(Token::API_SETTINGS);
+        $this->addApiToken(Token::API_EMAIL);
     }
 
     /**